Run as root and drop privs to non-root for most operations
In a cruel twist of fate, I recently discovered that OpenSMTPD errors
out on startup if its SSL certificate file has permissions more
permissive than 0755. This has been long-standing behavior in OpenSMTPD
that I didn't run into before (guess I don't restart smtpd very often).

The basis for le_renew_certs.sh was to run entirely as non-root which
required setting permissions on the SSL bundle file to 0775. This trips
OpenSMTPD's permissions check and causes the daemon to fail at startup.

Refactor the script to run as root and drop privs to non-root for
running the acme-tiny scripts and openssl(1). Root privs are maintained
for creating the cert bundle file so that the expected 0755 permissions
can be maintained.