New cmd line arg (-p) for specifying path to python
Trying to automatically find the right python version and the right path
to the interpreter is going to be a never-ending job. Leave the current
if/else ladder that does some rudimentary searches but allow user to
specify the exact path that's appropriate for their system thereby
(hopefully) taking care of this forever more.